+++ pybit/unified_trading.py
@@ -21,6 +21,7 @@ from ._v5_crypto_loan import CryptoLoanHTTP
 from ._v5_earn import EarnHTTP
 from ._v5_fiat import FiatHTTP
 from ._v5_rfq import RFQHTTP
+from ._v5_p2p import P2PHTTP
 from ._websocket_stream import _V5WebSocketManager
 from ._websocket_trading import _V5TradeWebSocketManager
 from ._v5_spread import (
@@ -64,6 +65,7 @@ class HTTP(
     FiatHTTP,
     RFQHTTP,
     RateLimitHTTP,
+    P2PHTTP,
     SpreadHTTP
 ):
     def __init__(self, **args):